fix: move `gamehud` to `interfaces` directory
